move file load code to a module
