+++ src/apps/throwdown/index.js
@@ -1,11 +1,7 @@
-import _ from 'lodash'; 
-
 import React from 'react';
 import { render } from 'react-dom';
 
 import { Provider } from 'react-redux';
-
-import Hjson from 'hjson';
 
 import store from './store/store';
 import observeStore from '@lib/observe-redux-store';
@@ -16,8 +12,7 @@ import Header from './components/transport-header/header.jsx';
 import HeaderPlaybackProgress from './components/playback-progress/header-progress.jsx';
 import Decks from './components/throwdown/decks.jsx';
 
-
-import throwdownActions from './components/throwdown/actions';
+import importThrowdownFile from './components/file-import/file-import';
 
 import './style/style.scss';
 
@@ -26,51 +21,8 @@ import './style/style.scss';
 
 const throwdownApp = new ThrowdownApp();
 
-
 /// -----------------------------------------------------------------------------------------------
 // load hard-coded test data
-
-function addThrowdownDeck( songSlug, throwdownData ) {
-  throwdownApp.importPatterns( songSlug, throwdownData.patterns );
-
-  store.dispatch( throwdownActions.addDeck( {
-    deckSlug: songSlug,
-  } ) ); 
-  
-  _.map( throwdownData.sections, ( section, key ) => {
-    store.dispatch( throwdownActions.addSection( {
-      deckSlug: songSlug,
-      slug: key, 
-      ...section
-    } ) );
-  } );
-
-  // trigger a random section
-  // const sectionSlugs = _.keys( throwdownData.sections );
-  // store.dispatch( throwdownActions.setDeckTriggeredSection( {
-  //   deckSlug: songSlug,
-  //   sectionSlug: _.sample( sectionSlugs )
-  // } ) );
-
-  // hard code build for testing
-  // store.dispatch( throwdownActions.setDeckTriggeredSection( {
-  //   deckSlug: 'test',
-  //   sectionSlug: 'build',
-  // } ) );
-}
-
-// const testSongFile = '/data/20190217--manas.hjson';
-// const testSongFile = 'data/20190306--sweets-from-a-stranger.hjson';
-// const testSongFile = 'data/20190306--its-not-real.hjson';
-
-function importThrowdownFile( deckSlug, file ) {
-  window.fetch( file )
-    .then( response => response.text() )
-    .then( text => {
-      const songData = Hjson.parse( text );
-      addThrowdownDeck( deckSlug, songData );
-    } );
-}
 
 importThrowdownFile( 'manas', '/data/20190217--manas.hjson' );
 // importThrowdownFile( 'noyu', '/data/20190325--noyu.hjson' );
@@ -79,7 +31,7 @@ importThrowdownFile( 'manas', '/data/20190217--manas.hjson' );
 // importThrowdownFile( 'maenyb', 'data/20190325--maenyb.hjson' );
 // importThrowdownFile( 'shedout', 'data/20190325--shedout.hjson' );
 // importThrowdownFile( 'mivova', 'data/20190325--mivova.hjson' );
-// importThrowdownFile( 'kufca', 'data/20190325--kufca.hjson' );
+importThrowdownFile( 'kufca', 'data/20190325--kufca.hjson' );
 
 /// -----------------------------------------------------------------------------------------------
 // bind sequencer/transport to store
